+++ CHANGELOG.md
@@ -10,8 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Config: Add `tools.float_win_config` and `replace_builtin_hover` to replace `tools.hover_actions`.
-  Would have to use the `tools.hover_actions` options if they exist.
+- Config: Add `tools.float_win_config` apply to floating Windows.
+  Move `border`, `max_width`, `max_height`, `auto_focus`
+  from `hover_actions` to `float_win_config`.
 
 ## [3.12.2] - 2024-01-07